requirements.txt
python-dotenv==1.0.0
requests==2.31.0
fastapi==0.104.1
uvicorn==0.23.2
fastembed==0.1.1
qdrant-client==1.5.4
pandas==2.1.1
numpy==1.26.0
python-multipart==0.0.6
starlette==0.27.0
pydantic==2.4.2
httpx==0.25.0